Use tabs to ensure rendering breakdown
