+++ docs/src/App.tsx
@@ -1,49 +1,42 @@
+import Tab from "@material-ui/core/Tab"
+import Tabs from "@material-ui/core/Tabs"
 import React, { Component } from "react"
 
-import { FPSClock, GameManager, loadReplay } from "../../src"
-import { GameBuilderOptions } from "../../src/builders/GameBuilder"
 import Main from "./components/Main"
 
+type ActiveTab = "viewer" | "other"
+
 interface State {
-  options?: GameBuilderOptions
+  tab: ActiveTab
 }
 
 class App extends Component<any, State> {
   constructor(props: any) {
     super(props)
-    this.state = {}
+    this.state = {
+      tab: "viewer",
+    }
   }
 
-  componentDidMount() {
-    const REPLAY_ID = "9944A36A11E987D3E286C1B524E68ECC"
-
-    // loadBuilderFromReplay(REPLAY_ID).then(gameManager => {
-    //   this.setState({ gameManager })
-    // })
-    loadReplay(REPLAY_ID).then(([replayData, replayMetadata]) => {
-      this.setState({
-        options: {
-          replayData,
-          replayMetadata,
-          clock: FPSClock.convertReplayToClock(replayData),
-        },
-      })
+  handleChange = (_: any, newTab: ActiveTab) => {
+    this.setState({
+      tab: newTab,
     })
   }
 
   render() {
-    const { options } = this.state
-
-    if (!options) {
-      return "Loading..."
-    }
-
+    const { tab } = this.state
     return (
       <div style={{ maxWidth: 900, width: "100%", margin: "0 auto" }}>
         <div>
           <h2>Welcome to React</h2>
         </div>
-        <Main options={options} />
+        <Tabs value={tab} onChange={this.handleChange}>
+          <Tab label="Viewer" value="viewer" />
+          <Tab label="Other" value="other" />
+        </Tabs>
+        {tab === "viewer" && <Main />}
+        {tab === "other" && <div>Other</div>}
       </div>
     )
   }
